requirements.txt
django==5.1.4
django-ninja==1.4.3
pydantic==2.11.5
pytest==8.3.5
pytest-django==4.11.1
pytest-cov==6.1.1
coverage==7.8.2

cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
injector==0.22.0
pycparser==2.22
pyjwt==2.10.1

uvicorn==0.34.2
fastapi==0.115.12
